+++ tests/Fixtures/CommandThatDoesNotHaveRunMethod.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Leapt\GitWrapper\Tests\Fixtures;
-
-final class CommandThatDoesNotHaveRunMethod
-{
-}