+++ tests/RepositoryTest.php
@@ -8,7 +8,6 @@ use Leapt\GitWrapper\Configuration;
 use Leapt\GitWrapper\Exception\GitRuntimeException;
 use Leapt\GitWrapper\Exception\InvalidGitRepositoryDirectoryException;
 use Leapt\GitWrapper\Repository;
-use Leapt\GitWrapper\Tests\Fixtures\CommandThatDoesNotHaveRunMethod;
 use Leapt\GitWrapper\Tests\Fixtures\CommandThatDoesNotImplementInterface;
 use PHPUnit\Framework\TestCase;
 
@@ -172,24 +171,10 @@ final class RepositoryTest extends TestCase
         new Repository($directory);
     }
 
-    public function testCreateCommandThrowsExceptionIfDoesNotHaveRunMethod(): void
+    public function testCreateCommandThrowsExceptionIfCommandClassDoesNotImplementInterface(): void
     {
         self::expectException(\RuntimeException::class);
-        Repository::createCommand(CommandThatDoesNotHaveRunMethod::class, 'test', 'test', false);
-    }
-
-    public function testCreateCommandThrowsDeprecationIfCommandClassDoesNotImplementInterface(): void
-    {
-        set_error_handler(
-            static function ($errno, $errstr) {
-                restore_error_handler();
-                throw new \Exception($errstr, $errno);
-            },
-            \E_ALL,
-        );
-
-        self::expectException(\Exception::class);
-        self::expectExceptionMessage('Passing a Command class that does not implement "Leapt\GitWrapper\CommandInterface" is deprecated.');
+        self::expectExceptionMessage('The Command class must implement the "Leapt\GitWrapper\CommandInterface" interface, the "Leapt\GitWrapper\Tests\Fixtures\CommandThatDoesNotImplementInterface" class does not.');
         Repository::createCommand(CommandThatDoesNotImplementInterface::class, 'test', 'test', false);
     }